backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
numpy-financial==1.0.0
python-dateutil==2.9.0
gunicorn==22.0.0
